Task 101

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -10,7 +10,7 @@
 
 void print_number(int n)
 {
-	unsigned double num;
+	unsigned int num;
 
 	if (n < 0)
 	{
